remove warning
git-svn-id: file:///home/mroda/Software/Genie/SVN/PhabGlobal/trunk@1769 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/EVGModules/VertexGenerator.cxx
@@ -110,7 +110,7 @@ void VertexGenerator::ProcessEventRecord(GHepRecord * evrec) const
         double rmax = 3*R;
         double dr   = R/40.;
         for(double r = 0; r < rmax; r+=dr) { 
-           ymax = TMath::Max(ymax, r*r * utils::nuclear::Density(r,A)); 
+           ymax = TMath::Max(ymax, r*r * utils::nuclear::Density(r,(int)A)); 
         }
         ymax *= 1.2;
         
@@ -132,7 +132,7 @@ void VertexGenerator::ProcessEventRecord(GHepRecord * evrec) const
 
            double r = rmax * rnd->RndFsi().Rndm();
            double t = ymax * rnd->RndFsi().Rndm();
-           double y = r*r * utils::nuclear::Density(r,A);
+           double y = r*r * utils::nuclear::Density(r,(int)A);
            if(y > ymax) {
                LOG("Vtx", pERROR)
                  << "y = " << y << " > ymax = " << ymax
